Sprint 01 04

+++ src/features/blogs/repositories/blogs.repository.ts
@@ -2,14 +2,53 @@ import { Blog } from '../types';
 import { BlogInputDto } from '../dto/blogInputDto';
 import { blogsCollection } from '../../../db/mongo.db';
 import { ObjectId, WithId } from 'mongodb';
+import { RepositoryNotFoundError } from '../../../core/errors/repository-not-found.error';
+import { BlogQueryInput } from '../routes/input/blog-query.input';
 
 export const blogsRepository = {
-  async findAll(): Promise<WithId<Blog>[]> {
-    return blogsCollection.find().toArray();
+  async findAll(
+    queryDto: BlogQueryInput,
+  ): Promise<{ items: WithId<Blog>[]; totalCount: number }> {
+    const {
+      pageSize,
+      pageNumber,
+      sortBy,
+      sortDirection,
+      searchWebsiteUrlTerm,
+      searchNameTerm,
+    } = queryDto;
+
+    const skip = (pageNumber - 1) * pageSize;
+    const filter: any = {};
+
+    if (searchNameTerm) {
+      filter.name = { $regex: searchNameTerm, $options: 'i' };
+    }
+
+    if (searchWebsiteUrlTerm) {
+      filter.websiteUrl = { $regex: searchWebsiteUrlTerm, $options: 'i' };
+    }
+
+    const items = await blogsCollection
+      .find(filter)
+      .sort({ [sortBy]: sortDirection })
+      .skip(skip)
+      .limit(pageSize)
+      .toArray();
+
+    const totalCount = await blogsCollection.countDocuments(filter);
+
+    return { items, totalCount };
   },
 
-  async findById(id: string): Promise<WithId<Blog> | null> {
-    return blogsCollection.findOne({ _id: new ObjectId(id) });
+  async findById(id: string): Promise<WithId<Blog>> {
+    const blog = await blogsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!blog) {
+      throw new RepositoryNotFoundError('Blog not exist');
+    }
+
+    return blog;
   },
 
   async create(newBlog: Blog): Promise<WithId<Blog>> {
@@ -18,6 +57,12 @@ export const blogsRepository = {
   },
 
   async update(id: string, dto: BlogInputDto): Promise<void> {
+    const blog = await blogsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!blog) {
+      throw new RepositoryNotFoundError('Blog not exist');
+    }
+
     const updateResult = await blogsCollection.updateOne(
       { _id: new ObjectId(id) },
       {
@@ -30,6 +75,7 @@ export const blogsRepository = {
     );
 
     if (updateResult.matchedCount < 1) {
+      //  TODO: заменить на DomainError если такая проверка есть в дз
       throw new Error('Blog not exist');
     }
 
@@ -37,11 +83,18 @@ export const blogsRepository = {
   },
 
   async delete(id: string): Promise<void> {
+    const blog = await blogsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!blog) {
+      throw new RepositoryNotFoundError('Blog not exist');
+    }
+
     const deleteResult = await blogsCollection.deleteOne({
       _id: new ObjectId(id),
     });
 
     if (deleteResult.deletedCount < 1) {
+      //  TODO: заменить на DomainError если такая проверка есть в дз
       throw new Error('Blog not exist');
     }
